Create Remove Nth Node From End of List.java
Remember the beauty of recursion. It can reverse the order of a list when the stack call returns. So simply we traverse the linked list till the end. Once we know that we have reached the end of list we set the flag to true indicating that we are now basically traversing backwards in the list.
Simply delete the nth node. Upvote if you liked this approach :)